backend/requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
cloudinary==1.40.0
python-dotenv==1.0.1
bcrypt==4.1.3
marshmallow==3.21.3
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.0.0
Werkzeug==3.0.3
gunicorn==22.0.0
